+++ packages/types/src/mesh-source-config.ts
@@ -30,12 +30,12 @@ export interface Handler {
   thrift?: ThriftHandler;
   tuql?: TuqlHandler;
   ContentfulHandler?: ContentfulHandler;
-  CrunchbaseHandler?: CrunchbaseHandler;
-  SalesforceHandler?: SalesforceHandler;
   SlackHandler?: SlackHandler;
   StripeHandler?: StripeHandler;
-  TwitterHandler?: TwitterHandler;
   WeatherbitHandler?: WeatherbitHandler;
+  CrunchbaseHandler?: CrunchbaseHandler;
+  SalesforceHandler?: SalesforceHandler;
+  TwitterHandler?: TwitterHandler;
   IPAPIHandler?: IPAPIHandler;
   [k: string]: unknown;
 }
@@ -99,10 +99,6 @@ export interface GraphQLHandler {
    */
   introspection?: string;
   /**
-   * Cache Introspection (Any of: GraphQLIntrospectionCachingOptions, Boolean)
-   */
-  cacheIntrospection?: GraphQLIntrospectionCachingOptions | boolean;
-  /**
    * Enable multipart/formdata in order to support file uploads
    */
   multipart?: boolean;
@@ -110,16 +106,6 @@ export interface GraphQLHandler {
    * Batch requests
    */
   batch?: boolean;
-}
-export interface GraphQLIntrospectionCachingOptions {
-  /**
-   * Time to live of introspection cache
-   */
-  ttl?: number;
-  /**
-   * Path to Introspection JSON File
-   */
-  path?: string;
 }
 /**
  * Handler for gRPC and Protobuf schemas
@@ -430,16 +416,6 @@ export interface Neo4JHandler {
    * Provide GraphQL Type Definitions instead of inferring
    */
   typeDefs?: string;
-  /**
-   * Cache Introspection (Any of: Neo4jIntrospectionCachingOptions, Boolean)
-   */
-  cacheIntrospection?: Neo4JIntrospectionCachingOptions | boolean;
-}
-export interface Neo4JIntrospectionCachingOptions {
-  /**
-   * Time to live of introspection cache
-   */
-  ttl?: number;
 }
 /**
  * Handler for OData
@@ -481,7 +457,12 @@ export interface OpenapiHandler {
   /**
    * A pointer to your API source - could be a local file, remote file or url endpoint
    */
-  source: string;
+  source:
+    | {
+        [k: string]: unknown;
+      }
+    | string
+    | unknown[];
   /**
    * Format of the source file (Allowed values: json, yaml)
    */
@@ -590,10 +571,6 @@ export interface PostGraphileHandler {
         [k: string]: unknown;
       }
     | string;
-  /**
-   * Cache Introspection (Any of: GraphQLIntrospectionCachingOptions, Boolean)
-   */
-  cacheIntrospection?: GraphQLIntrospectionCachingOptions | boolean;
   /**
    * Enable GraphQL websocket transport support for subscriptions (default: true)
    */
@@ -732,16 +709,57 @@ export interface TuqlHandler {
    */
   infile?: string;
 }
+/**
+ * API-first content platform to build digital experiences
+ */
 export interface ContentfulHandler {
   /**
    * Authentication token
    */
   token: string;
   /**
-   * A endpoint of your Contentful GraphQL API
+   * (Deprecated) Full endpoint including space ID and environment. More info: https://www.contentful.com/developers/docs/references/graphql/#/introduction/basic-api-information
    */
-  endpoint: string;
+  endpoint?: string;
+  /**
+   * Your space ID. Wonder how to find it? Check https://www.contentful.com/help/find-space-id/
+   */
+  space?: string;
+  /**
+   * Your environment ID. What is it? Check https://www.contentful.com/faq/environments/
+   */
+  environment?: string;
 }
+/**
+ * A proprietary business communication platform
+ */
+export interface SlackHandler {
+  /**
+   * Slack API access token
+   */
+  token: string;
+}
+/**
+ * Online payment processing for internet businesses
+ */
+export interface StripeHandler {
+  /**
+   * Stripe secret key
+   */
+  token: string;
+}
+/**
+ * Weather API to instantly access real-time and historical data
+ */
+export interface WeatherbitHandler {
+  /**
+   * Weatherbit secret key
+   */
+  token: string;
+}
+/**
+ * A platform for finding business information about private and public companies
+ */
 export interface CrunchbaseHandler {
   /**
    * Authentication API Key
@@ -758,34 +776,18 @@ export interface SalesforceHandler {
    */
   token?: string;
 }
-export interface SlackHandler {
-  /**
-   * Slack API access token
-   */
-  token: string;
-}
-export interface StripeHandler {
-  /**
-   * Authentication token
-   */
-  token: string;
-}
+/**
+ * An American microblogging and social networking service on which users post and interact with messages known as 'tweets'
+ */
 export interface TwitterHandler {
   /**
-   * Authorization header. Set up the following variables in order to run each request (depending on the authentication type used by the request you are sending):
-   *
-   * |Name|Description|
-   * |---|---|
-   * |`consumer_key`|Your consumer key|
-   * |`consumer_secret`|Your consumer secret|
-   * |`access_token`|Your access token|
-   * |`token_secret`|Your access token secret|
-   * |`bearer_token`|Your bearer token|
-   *
+   * Your bearer token (wihout 'Bearer ' substring)
    */
-  authorization: string;
+  token: string;
 }
-export interface WeatherbitHandler {}
+/**
+ * IP Geolocation API
+ */
 export interface IPAPIHandler {}
 export interface Transform {
   /**